telegrambot/requirements.txt
python-telegram-bot
aiomqtt
certifi
environs
aiomysql